+++ h2integrate/control/control_strategies/storage/heuristic_pyomo_controller.py
@@ -6,7 +6,7 @@ from attrs import field, define
 
 from h2integrate.core.utilities import merge_shared_inputs
 from h2integrate.core.validators import range_val_or_none
-from h2integrate.control.control_strategies.pyomo_controller_baseclass import (
+from h2integrate.control.control_strategies.pyomo_storage_controller_baseclass import (
     PyomoStorageControllerBaseClass,
     PyomoStorageControllerBaseConfig,
 )
